requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
typer>=0.9.0
tqdm>=4.65.0
pyyaml>=6.0

# LLM providers
openai>=1.0.0
httpx>=0.24.0

# Data analysis and visualization
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0
statsmodels>=0.14.0

# Testing
pytest>=7.3.0
pytest-asyncio>=0.21.0
